requirements.txt
Flask==2.3.2
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
beautifulsoup4==4.12.2
Werkzeug==2.3.6